refactor: 优化项目结构
