+++ src/views/maint/database/index.vue
@@ -67,8 +67,8 @@
 </template>
 
 <script>
-import crudDatabase from '@/api/mnt/database'
-import { testDbConnect } from '@/api/mnt/connect'
+import crudDatabase from '@/api/maint/database'
+import { testDbConnect } from '@/api/maint/connect'
 import eForm from './execute'
 import CRUD, { presenter, header, form, crud } from '@crud/crud'
 import rrOperation from '@crud/RR.operation'
