+++ src/views/maint/server/index.vue
@@ -59,8 +59,8 @@
 
 <script>
 
-import crudServer from '@/api/mnt/serverDeploy'
-import { testServerConnect } from '@/api/mnt/connect'
+import crudServer from '@/api/maint/serverDeploy'
+import { testServerConnect } from '@/api/maint/connect'
 import { validateIP } from '@/utils/validate'
 import CRUD, { presenter, header, form, crud } from '@crud/crud'
 import rrOperation from '@crud/RR.operation'
